Added ACL user checking to CMSysBot.
Now, for any user to be able to connect to a section, the username used for the connection must be on the allowed_users list, defined on the config.json file.
If the list is empty, all users will be able to connect to that specific section.

NOTE: The assigned user permissions will be propagated to all subsections. That means that if any user has access to a section, they will also have access to all subsections below