🚑(Plugin) Plugin no longer includes all computers, ignoring filtering
The function `run` has been modified to use the `get_included_computers()`
function from the `Computers` class, and not the `get_computers()` one.

This means that now the Plugins should only execute on the filtered
computers, as expected.

Also, now the variable `MACS_LIST` only returns the MAC from the
included computers.

